+++ Sources/Core/State/PlayingState.swift
@@ -125,8 +125,10 @@ final class PlayingState: PlayerState {
     private func setupPlaybackObservingCallback() {
         itemPlaybackObservingService.onPlaybackStalled = { [weak self] in self?.redirectToWaitingForNetworkState() }
         itemPlaybackObservingService.onFailedToPlayToEndTime = { [weak self] in self?.redirectToWaitingForNetworkState() }
-        itemPlaybackObservingService.onPlayToEndTime = { [weak self] in
-            self?.context.nowPlaying.overrideInfoCenter(for: MPNowPlayingInfoPropertyElapsedPlaybackTime, value: NSNumber(value: 0))
+        itemPlaybackObservingService.onPlayToEndTime = { [weak self, context] in
+            context.delegate?.playerContext(didItemPlayToEndTime: context.currentTime)
+            context.plugins.forEach { $0.didItemPlayToEndTime(endTime: context.currentTime) }
+            context.nowPlaying.overrideInfoCenter(for: MPNowPlayingInfoPropertyElapsedPlaybackTime, value: NSNumber(value: 0))
             self?.redirectToStoppedState()
         }
     }